Added BasePopup.
